Update `CpuInfo::get_info` to reduce number of needless allocations
When creating an `HashMap`, initially its size is small and adding elements
into would induce re-allocations. This behaviour is ok in the general case,
but if we are able to pre-determine the final size of the map then we would
avoid not needed allocations and performances would be improved.
NOTE: When building an `HashMap` from an `Iterator` uses it uses the
      `Iterator::size_hint` API to pre-allocate the `HashMap`, hence
      re-writing the method in a more rust-like apporach leds to better
      performances.