Update github-actions configuration to run all the github actions up to completion
By disabling fail-fast in the actions matrix we ensure that all the jobs defined by the
matrix are executed up to their completion.
This is useful because while running checks or tests on multiple toolchain versions it is
useful to have the report of all the versions and not and not stop the tests as soon as
1 toolchain version failed.
NOTE: This does not alter the fact that the overall actions run will be marked as failed.